requirements.txt
pyyaml
pytest
pytest-bdd
pytest-xdist
appium-python-client
allure-pytest